HDDS-12750. Move StorageTypeProto from ScmServerDatanodeHeartbeatProtocol.proto to hdds.proto (#8208)
(cherry picked from commit bba8a6783114bb2e9bd05755232e561274b74714)
