+++ hadoop-hdds/interface-server/src/main/proto/ScmServerDatanodeHeartbeatProtocol.proto
@@ -193,16 +193,6 @@ message MetadataStorageReportProto {
   optional bool failed = 6 [default = false];
 }
 
-/**
- * Types of recognized storage media.
- */
-enum StorageTypeProto {
-  DISK = 1;
-  SSD = 2;
-  ARCHIVE = 3;
-  RAM_DISK = 4;
-  PROVIDED = 5;
-}
 
 message ContainerReportsProto {
   repeated ContainerReplicaProto reports = 1;
