+++ hadoop-ozone/integration-test/src/test/java/org/apache/hadoop/ozone/om/TestBucketLayoutWithOlderClient.java
@@ -21,6 +21,7 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import java.util.UUID;
+import org.apache.hadoop.hdds.protocol.proto.HddsProtos.StorageTypeProto;
 import org.apache.hadoop.hdds.utils.IOUtils;
 import org.apache.hadoop.ozone.ClientVersion;
 import org.apache.hadoop.ozone.MiniOzoneCluster;
@@ -88,7 +89,7 @@ public abstract class TestBucketLayoutWithOlderClient implements NonHATests.Test
                         OzoneManagerProtocolProtos.BucketInfo.newBuilder()
                             .setVolumeName(volumeName).setBucketName(buckName)
                             .setIsVersionEnabled(false).setStorageType(
-                            OzoneManagerProtocolProtos.StorageTypeProto.DISK)
+                            StorageTypeProto.DISK)
                             .build())
                     .build()).build();
     createBucketReq = createBucketReq.toBuilder()
